+++ tests/type_tests.rs
@@ -15,8 +15,8 @@ fn create_test_model(id: &str) -> all_types::ActiveModel {
         string_nullable: Set(Some("nullable string".to_string())),
         int64_val: Set(9223372036854775807i64),
         int64_nullable: Set(Some(-9223372036854775808i64)),
-        int32_val: Set(2147483647),
-        int32_nullable: Set(Some(-2147483648)),
+        int32_val: Set(2147483647_i64),
+        int32_nullable: Set(Some(-2147483648_i64)),
         float64_val: Set(3.141592653589793),
         float64_nullable: Set(Some(-1.7976931348623157e308)),
         float32_val: Set(3.14159f64),
@@ -41,7 +41,7 @@ fn create_test_model_with_nulls(id: &str) -> all_types::ActiveModel {
         string_nullable: Set(None),
         int64_val: Set(i64::MIN),
         int64_nullable: Set(None),
-        int32_val: Set(0),
+        int32_val: Set(0_i64),
         int32_nullable: Set(None),
         float64_val: Set(0.0),
         float64_nullable: Set(None),
@@ -171,22 +171,22 @@ mod integer_type_tests {
         let id = uuid::Uuid::new_v4().to_string();
 
         let model = all_types::ActiveModel {
-            int32_val: Set(i32::MAX),
-            int32_nullable: Set(Some(i32::MIN)),
+            int32_val: Set(i32::MAX as i64),
+            int32_nullable: Set(Some(i32::MIN as i64)),
             ..create_test_model(&id)
         };
 
         let inserted = model.insert(&db).await.expect("Insert failed");
-        assert_eq!(inserted.int32_val, i32::MAX);
-        assert_eq!(inserted.int32_nullable, Some(i32::MIN));
+        assert_eq!(inserted.int32_val, i32::MAX as i64);
+        assert_eq!(inserted.int32_nullable, Some(i32::MIN as i64));
 
         let selected = all_types::Entity::find_by_id(&id)
             .one(&db)
             .await
             .expect("Select failed")
             .expect("Entity not found");
-        assert_eq!(selected.int32_val, i32::MAX);
-        assert_eq!(selected.int32_nullable, Some(i32::MIN));
+        assert_eq!(selected.int32_val, i32::MAX as i64);
+        assert_eq!(selected.int32_nullable, Some(i32::MIN as i64));
     }
 
     #[tokio::test]
@@ -198,16 +198,16 @@ mod integer_type_tests {
         let model = all_types::ActiveModel {
             int64_val: Set(i64::MIN),
             int64_nullable: Set(Some(i64::MAX)),
-            int32_val: Set(-999),
-            int32_nullable: Set(Some(0)),
+            int32_val: Set(-999_i64),
+            int32_nullable: Set(Some(0_i64)),
             ..create_test_model(&id)
         };
 
         let inserted = model.insert(&db).await.expect("Insert failed");
         assert_eq!(inserted.int64_val, i64::MIN);
         assert_eq!(inserted.int64_nullable, Some(i64::MAX));
-        assert_eq!(inserted.int32_val, -999);
-        assert_eq!(inserted.int32_nullable, Some(0));
+        assert_eq!(inserted.int32_val, -999_i64);
+        assert_eq!(inserted.int32_nullable, Some(0_i64));
 
         let selected = all_types::Entity::find_by_id(&id)
             .one(&db)
@@ -216,8 +216,8 @@ mod integer_type_tests {
             .expect("Entity not found");
         assert_eq!(selected.int64_val, i64::MIN);
         assert_eq!(selected.int64_nullable, Some(i64::MAX));
-        assert_eq!(selected.int32_val, -999);
-        assert_eq!(selected.int32_nullable, Some(0));
+        assert_eq!(selected.int32_val, -999_i64);
+        assert_eq!(selected.int32_nullable, Some(0_i64));
     }
 
     #[tokio::test]
